# ActionCable: optimise logger.debug calling